+- add patches.fixes/linux-post-2.6.3-20040220 +  update to linus latest, usb, ia64, xeon64, i4l, ide fixes
suse-commit: a7cbdb4e4e99832b3a154ba4c89ec0d6e68d821b

+++ drivers/isdn/hisax/callc.c
@@ -918,7 +918,7 @@ static void stat_redir_result(struct IsdnCardState *cs, int chan, ulong result)
 	ic.driver = cs->myid;
 	ic.command = ISDN_STAT_REDIR;
 	ic.arg = chan; 
-	(ulong)(ic.parm.num[0]) = result;
+	ic.parm.num[0] = result;
 	cs->iif.statcallb(&ic);
 } /* stat_redir_result */
